refactor(shared): flatten sidebar structure, update shared UI
