+++ src/shared/ui/sidebar/const.ts
